+++ drizzle.config.ts
@@ -1,7 +1,12 @@
 import { defineConfig } from 'drizzle-kit';
 
 export default defineConfig({
-  schema: './src/lib/db/schema/*',
+  schema: [
+    './src/lib/db/schema/products.ts',
+    './src/lib/db/schema/users.ts',
+    './src/lib/db/schema/kanban.ts',
+    './src/lib/db/schema/notifications.ts'
+  ],
   out: './src/lib/db/migrations',
   dialect: 'postgresql',
   dbCredentials: {