+++ src/lib/db/migrations/0003_cheerful_rumiko_fujikawa.sql
@@ -1,0 +1,10 @@
+CREATE TYPE "public"."notification_status" AS ENUM('unread', 'read', 'archived');--> statement-breakpoint
+CREATE TABLE "notifications" (
+	"id" serial PRIMARY KEY NOT NULL,
+	"title" text NOT NULL,
+	"body" text NOT NULL,
+	"status" "notification_status" DEFAULT 'unread' NOT NULL,
+	"actions" jsonb,
+	"created_at" timestamp DEFAULT now() NOT NULL,
+	"updated_at" timestamp DEFAULT now() NOT NULL
+);