+++ src/lib/db/schema/index.ts
@@ -1,3 +1,4 @@
 export * from './products';
 export * from './users';
 export * from './kanban';
+export * from './notifications';